+++ Day07-We_Still_Need_To_Keep_An_Eye_On_This/Day-07-proposed/Day-07-proposed.ino
@@ -25,10 +25,10 @@
 #include "Arduino.h"
 
 // Our photoresistor will give us a reading of the current light level on this analog pin
-const uint8_t PHOTORESISTOR_PIN = A0;  // Photoresistor analog pin
+const byte PHOTORESISTOR_PIN = A0;  // Photoresistor analog pin
 
 // Use an unsigned, 16 bit value (0 - 65535) for maximum battery capacity
-const uint16_t BATTERY_CAPACITY = 50000;  // Maximum battery capacity
+const unsigned int BATTERY_CAPACITY = 50000;  // Maximum battery capacity
 
 void setup() {
   pinMode(PHOTORESISTOR_PIN, INPUT);  // Input current light level from photoresistor
@@ -45,7 +45,7 @@ void setup() {
 void loop() {
   // Use static because we need this variable to maintain it's value across
   // multiple loop() runs.
-  static uint16_t battery_level = 0;  // Current battery charge level (set to 0 first time)
+  static unsigned int battery_level = 0;  // Current battery charge level (set to 0 first time)
 
   // Display battery information on the Serial Monitor
   if (battery_level < BATTERY_CAPACITY) {  // If battery is not fully charged..
